divide note details component into smaller components
